Single station add
